fix: update tests to use BigDecimal for fare
- Update RideRequestTestBuilder to use BigDecimal for fare
- Update test assertions to compare BigDecimal values
- Fix mock return types for calculateFare method

+++ ubercloneapp/src/test/java/com/github/k2ocabhinav/ubercloneapp/services/impl/DriverServiceImplTest.java
@@ -103,7 +103,7 @@ class DriverServiceImplTest {
                 .driver(testDriver)
                 .rideStatus(RideStatus.CONFIRMED)
                 .otp("1234")
-                .fare(BigDecimal.valueOf(100.0))
+                .fare(100.0)
                 .paymentMethod(PaymentMethod.CASH)
                 .build();
     }
